1.更新前端页面，完善了6个页面1.计算图可视化 2.学习率和迭代次数可视化 3.梯度下降可视化；2.优化可视化图片大小
