+++ LearnTorch_ALL/LearnTorch/__init__.py
@@ -1,5 +1,6 @@
 from LearnTorch.VariableFunction import Variable
 from LearnTorch.VariableFunction import Function
+from LearnTorch.VariableFunction import Parameter
 from LearnTorch.VariableFunction import using_config
 from LearnTorch.VariableFunction import no_grad
 from LearnTorch.VariableFunction import as_array
