--- VirtualAssistant/Pages/Explore.py
@@ -1,5 +1,0 @@
-import streamlit as st
-
-st.header("探索属于你的虚拟助手")
-
-
