+++ VirtualAssistant/README.md
@@ -1,3 +1,3 @@
 大致功能已经完成 外观和小bug还可以优化
 运行方式：
-anaconda环境下 终端输入streamlit run HomePages.py
+anaconda环境下 终端输入streamlit run Stream_litapps.py
